billbrain-unified-v2/billbrain-unified/requirements.txt
# BillBrain Unified - Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25

# LLM
ollama==0.1.6

# Gmail API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
